+++ backend/app/text_shape_to_sound/views.py
@@ -29,7 +29,7 @@ Example:
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from django.shortcuts import render
-from app.data_processing import text_shape_to_sound
+from app.text_shape_to_sound import text_shape_to_sound
 from app.synthesis.audio_encoding import audio_samples_to_wav_base64
 
 def index(request):
